refactor(ui): move charts folder into component, remove console.debug, improve xAxisData function
