+++ ui/src/container/models/Details/binary-classification/current/model-quality/charts.jsx
@@ -1,5 +1,5 @@
-import { CHART_COLOR } from '@Container/models/Details/charts/common-chart-options';
-import LineChart from '@Container/models/Details/charts/line-chart';
+import { CHART_COLOR } from '@Helpers/common-chart-options';
+import LineChart from '@Components/charts/line-chart';
 import { MODEL_QUALITY_FIELD } from '@Container/models/Details/constants';
 import { useGetCurrentModelQualityQueryWithPolling } from '@Src/store/state/models/polling-hook';
 import { modelsApiSlice } from '@State/models/api';
