+++ ui/src/container/models/Details/binary-classification/reference/model-quality/index.jsx
@@ -1,5 +1,5 @@
 import JobStatus from '@Components/JobStatus';
-import ConfusionMatrix from '@Container/models/Details/charts/confusion-matrix-chart';
+import ConfusionMatrix from '@Components/charts/confusion-matrix-chart';
 import { MODEL_QUALITY_FIELD } from '@Container/models/Details/constants';
 import { JOB_STATUS } from '@Src/constants';
 import { useGetReferenceModelQualityQueryWithPolling } from '@State/models/polling-hook';
@@ -7,7 +7,7 @@ import {
   Board, DataTable, SectionTitle, Spinner,
 } from '@radicalbit/radicalbit-design-system';
 import { memo } from 'react';
-import { CHART_COLOR } from '@Container/models/Details/charts/common-chart-options';
+import { CHART_COLOR } from '@Helpers/common-chart-options';
 import columns from './columns';
 
 function BinaryClassificationModelQualityMetrics() {
