+++ ui/src/container/models/Details/multi-classification/current/data-drift/use-get-filtered-features.js
@@ -28,8 +28,6 @@ export default () => {
       const isNumerical = isNumericalSelected && type === DRIFT_TEST_ENUM.KS;
       const isCategorical = isCategoricalSelected && type === DRIFT_TEST_ENUM.CHI2;
 
-      console.debug(type, isNumerical, isCategorical);
-
       return isNumerical || isCategorical;
     });
 };
