+++ ui/src/container/models/Details/multi-classification/current/model-quality/index.jsx
@@ -1,6 +1,6 @@
 import SomethingWentWrong from '@Components/ErrorPage/something-went-wrong';
 import JobStatus from '@Components/JobStatus';
-import ConfusionMatrix from '@Container/models/Details/charts/confusion-matrix-chart';
+import ConfusionMatrix from '@Components/charts/confusion-matrix-chart';
 import { MODEL_QUALITY_FIELD } from '@Container/models/Details/constants';
 import { JOB_STATUS } from '@Src/constants';
 import { modelsApiSlice } from '@State/models/api';
@@ -10,7 +10,7 @@ import {
 } from '@radicalbit/radicalbit-design-system';
 import { memo } from 'react';
 import { useParams } from 'react-router';
-import { CHART_COLOR } from '@Container/models/Details/charts/common-chart-options';
+import { CHART_COLOR } from '@Helpers/common-chart-options';
 import {
   AccuracyChart,
   AreaUnderPrChart,
