+++ ui/src/container/models/Details/multi-classification/reference/data-quality/data-point-distribution/options.js
@@ -1,6 +1,6 @@
-import { CHART_COLOR } from '@Container/models/Details/charts/common-chart-options';
+import { CHART_COLOR } from '@Helpers/common-chart-options';
 import { numberFormatter } from '@Src/constants';
-import * as commonChartOptions from '@Container/models/Details/charts/common-chart-options';
+import * as commonChartOptions from '@Helpers/common-chart-options';
 
 export default function chartOptions(title, dataset) {
   const yAxisLabel = dataset.map(({ name }) => name);
