+++ ui/src/container/models/Details/multi-classification/reference/data-quality/data-quality-list/numerical/chart/index.jsx
@@ -1,4 +1,4 @@
-import { CHART_COLOR } from '@Container/models/Details/charts/common-chart-options';
+import { CHART_COLOR } from '@Helpers/common-chart-options';
 import ReactEchartsCore from 'echarts-for-react/lib/core';
 import { BarChart } from 'echarts/charts';
 import {
