+++ ui/src/container/models/Details/multi-classification/reference/data-quality/data-quality-list/numerical/chart/options.js
@@ -1,10 +1,12 @@
 import { numberFormatter } from '@Src/constants';
-import * as commonChartOptions from '@Container/models/Details/charts/common-chart-options';
+import * as commonChartOptions from '@Helpers/common-chart-options';
 
 export default function chartOptions(dataset, color) {
-  const dataFormatted = dataset.buckets.map((value) => numberFormatter().format(value));
-  const lastElementData = dataFormatted.pop();
-  const xAxisData = dataFormatted.map((el, idx) => `[${dataFormatted[idx]}${(idx < dataFormatted.length - 1) ? `-${dataFormatted[idx + 1]})` : (idx === dataFormatted.length - 1) ? `-${lastElementData}]` : ''} `);
+  const { length, [length - 1]: last, ...rest } = dataset.buckets.map((value) => numberFormatter().format(value));
+
+  const values = Object.values(rest);
+
+  const xAxisData = values.map((el, idx) => `[${el}${(idx < values.length - 1) ? `-${values[idx + 1]})` : (idx === values.length - 1) ? `-${last}]` : ''} `);
 
   const options = {
     ...commonChartOptions.gridOptions.barChart(),
